woocommerce module reorganize & animation added
